chore: cleanup, fix: the tests finally pass
