+++ src/common/utils/shared.rs
@@ -5,8 +5,7 @@ use std::{iter, str::from_utf8};
 use strip_ansi_escapes::strip;
 
 use colors_transform::{Color, Rgb};
-use xrdb::Colors;
-use zellij_tile::data::{Palette, PaletteSource, Theme};
+use zellij_tile::data::{Theme};
 
 fn ansi_len(s: &str) -> usize {
     from_utf8(&strip(s.as_bytes()).unwrap())
@@ -40,7 +39,7 @@ pub mod colors {
     pub const GRAY: (u8, u8, u8) = (68, 68, 68);
     pub const BRIGHT_GRAY: (u8, u8, u8) = (138, 138, 138);
     pub const RED: (u8, u8, u8) = (135, 0, 0);
-    pub const ORANGE: (u8, u8, u8) = (215,95,0);
+    pub const ORANGE: (u8, u8, u8) = (215, 95, 0);
     pub const BLACK: (u8, u8, u8) = (0, 0, 0);
 }
 
@@ -64,46 +63,4 @@ pub fn detect_theme(bg: (u8, u8, u8)) -> Theme {
         true => Theme::Light,
         false => Theme::Dark,
     }
-}
-
-pub fn load_palette() -> Palette {
-    let palette = match Colors::new("xresources") {
-        Some(palette) => {
-            let fg = hex_to_rgb(&palette.fg);
-            let bg = hex_to_rgb(&palette.bg);
-            let colors: Vec<(u8, u8, u8)> = palette.colors.iter().map(|c| hex_to_rgb(c)).collect();
-            let theme = detect_theme(bg);
-            Palette {
-                source: PaletteSource::Xresources,
-                theme,
-                fg,
-                bg,
-                black: colors[0],
-                red: colors[1],
-                green: colors[2],
-                yellow: colors[3],
-                blue: colors[4],
-                magenta: colors[5],
-                cyan: colors[6],
-                white: colors[7],
-                orange: colors[9]
-            }
-        }
-        None => Palette {
-            source: PaletteSource::Default,
-            theme: Theme::Dark,
-            fg: colors::BRIGHT_GRAY,
-            bg: colors::GRAY,
-            black: colors::BLACK,
-            red: colors::RED,
-            green: colors::GREEN,
-            yellow: colors::GRAY,
-            blue: colors::GRAY,
-            magenta: colors::GRAY,
-            cyan: colors::GRAY,
-            white: colors::WHITE,
-            orange: colors::ORANGE
-        },
-    };
-    palette
 }
