+++ zellij-tile/Cargo.toml
@@ -7,6 +7,7 @@ description = "A small client-side library for writing Zellij plugins"
 license = "MIT"
 
 [dependencies]
+ansi_term = "0.12.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 strum = "0.20.0"
