+++ zellij-tile/src/data.rs
@@ -67,13 +67,13 @@ impl Default for InputMode {
     }
 }
 
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub enum Theme {
     Light,
     Dark,
 }
 
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub enum PaletteSource {
     Default,
     Xresources,
@@ -84,11 +84,11 @@ pub mod colors {
     pub const GRAY: (u8, u8, u8) = (68, 68, 68);
     pub const BRIGHT_GRAY: (u8, u8, u8) = (138, 138, 138);
     pub const RED: (u8, u8, u8) = (135, 0, 0);
-    pub const ORANGE: (u8, u8, u8) = (215,95,0);
+    pub const ORANGE: (u8, u8, u8) = (215, 95, 0);
     pub const BLACK: (u8, u8, u8) = (0, 0, 0);
 }
 
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub struct Palette {
     pub source: PaletteSource,
     pub theme: Theme,
@@ -102,7 +102,7 @@ pub struct Palette {
     pub magenta: (u8, u8, u8),
     pub cyan: (u8, u8, u8),
     pub white: (u8, u8, u8),
-    pub orange: (u8, u8, u8)
+    pub orange: (u8, u8, u8),
 }
 
 impl Default for Palette {
@@ -120,7 +120,7 @@ impl Default for Palette {
             magenta: colors::GRAY,
             cyan: colors::GRAY,
             white: colors::WHITE,
-            orange: colors::ORANGE
+            orange: colors::ORANGE,
         }
     }
 }
